QPPCT-715: loop in a few more smells
