+++ converter/src/main/java/gov/cms/qpp/conversion/model/error/QppValidationException.java
@@ -1,7 +1,7 @@
 package gov.cms.qpp.conversion.model.error;
 
 
-import gov.cms.qpp.conversion.Converter;
+import gov.cms.qpp.conversion.ConversionReport;
 
 /**
  * An exception thrown when causal exceptions arise during QPP validation.
@@ -15,7 +15,7 @@ public class QppValidationException extends TransformException {
 	 * @param cause A Throwable that caused this exception to occur.
 	 * @param report A report on the detail of the conversion.
 	 */
-	public QppValidationException(String message, Throwable cause, Converter.ConversionReport report) {
+	public QppValidationException(String message, Throwable cause, ConversionReport report) {
 		super(message, cause, report);
 	}
